lambda/requirements.txt
biopython==1.81
boto3==1.26.137